+++ internal/api/core/marketplace_paths.go
@@ -1,3 +1,7 @@
 package core
 
-const marketplaceExtensionsInstalledPath = "/marketplace/extensions"
+const (
+	marketplaceExtensionsInstalledPath = "/marketplace/extensions"
+	marketplaceMCPsInstalledPath       = "/marketplace/mcps"
+	marketplaceSkillsInstalledPath     = "/marketplace/skills"
+)
