+++ web/src/routes/_app/marketplace.$kind.$entryId.tsx
@@ -16,7 +16,7 @@ export const Route = createFileRoute("/_app/marketplace/$kind/$entryId")({
       parentCrumb: isMarketplaceKind(params.kind)
         ? {
             label: MARKETPLACE_KIND_LABEL[params.kind],
-            search: search.tab === "market" ? { tab: "market" } : {},
+            search: { scope: search.scope, tab: search.tab, workspace_id: search.workspace_id },
             to: `/marketplace/${marketplaceRouteKindFor(params.kind)}`,
           }
         : undefined,
