+++ web/src/systems/marketplace/components/marketplace-detail-manage-state.tsx
@@ -4,6 +4,52 @@ import { Button, Spinner } from "@compozy/ui";
 
 import { MarketplaceDetailRailCard } from "./marketplace-detail-shell";
 
+interface MarketplaceDetailManageFallbackBodyProps {
+  error?: Error | null;
+  errorMessage: string;
+  isLoading?: boolean;
+  loadingMessage: string;
+  onRetry: () => void;
+  testId: string;
+}
+
+/** Shared Manage loading/error body while the installed record is unavailable. */
+function MarketplaceDetailManageFallbackBody({
+  error,
+  errorMessage,
+  isLoading = false,
+  loadingMessage,
+  onRetry,
+  testId,
+}: MarketplaceDetailManageFallbackBodyProps) {
+  return (
+    <div
+      className="flex flex-col gap-2 px-3.5 py-1.5"
+      data-testid={testId}
+      role={isLoading ? "status" : "alert"}
+    >
+      <div className="flex items-center gap-2 text-small-body font-medium text-fg">
+        {isLoading ? (
+          <Spinner aria-hidden="true" className="size-3.5 text-subtle" />
+        ) : (
+          <AlertCircle aria-hidden="true" className="size-3.5 text-danger" />
+        )}
+        {isLoading ? loadingMessage : errorMessage}
+      </div>
+      {!isLoading && error?.message ? (
+        <p className="text-form-hint text-muted">{error.message}</p>
+      ) : null}
+      {!isLoading ? (
+        <div className="pb-1">
+          <Button onClick={onRetry} size="sm" type="button" variant="outline">
+            Retry management
+          </Button>
+        </div>
+      ) : null}
+    </div>
+  );
+}
+
 interface MarketplaceDetailManageFallbackCardProps {
   error?: Error | null;
   isLoading?: boolean;
@@ -22,33 +68,17 @@ function MarketplaceDetailManageFallbackCard({
 }: MarketplaceDetailManageFallbackCardProps) {
   return (
     <MarketplaceDetailRailCard icon={SlidersHorizontal} title="Manage">
-      <div
-        className="flex flex-col gap-2 px-3.5 py-1.5"
-        data-testid={testId}
-        role={isLoading ? "status" : "alert"}
-      >
-        <div className="flex items-center gap-2 text-small-body font-medium text-fg">
-          {isLoading ? (
-            <Spinner aria-hidden="true" className="size-3.5 text-subtle" />
-          ) : (
-            <AlertCircle aria-hidden="true" className="size-3.5 text-danger" />
-          )}
-          {isLoading ? `Loading ${label} management…` : `${label} management could not be loaded.`}
-        </div>
-        {!isLoading && error?.message ? (
-          <p className="text-form-hint text-muted">{error.message}</p>
-        ) : null}
-        {!isLoading ? (
-          <div className="pb-1">
-            <Button onClick={onRetry} size="sm" type="button" variant="outline">
-              Retry management
-            </Button>
-          </div>
-        ) : null}
-      </div>
+      <MarketplaceDetailManageFallbackBody
+        error={error}
+        errorMessage={`${label} management could not be loaded.`}
+        isLoading={isLoading}
+        loadingMessage={`Loading ${label} management…`}
+        onRetry={onRetry}
+        testId={testId}
+      />
     </MarketplaceDetailRailCard>
   );
 }
 
-export { MarketplaceDetailManageFallbackCard };
-export type { MarketplaceDetailManageFallbackCardProps };
+export { MarketplaceDetailManageFallbackBody, MarketplaceDetailManageFallbackCard };
+export type { MarketplaceDetailManageFallbackBodyProps, MarketplaceDetailManageFallbackCardProps };
